Сделал предварительную вёрстку лэйаута страницы
