+++ src/components/app-header/app-header.module.css
@@ -1,3 +1,9 @@
 .header {
-  flex-basis: 100%
+  width: 100%;
+  min-height: 50px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  background-color: teal;
+  color: white;
 }
