+++ src/components/burger-constructor/burger-constructor.module.css
@@ -1,3 +1,9 @@
 .section {
-  flex-basis: 50%;
+  display: flex;
+  flex: 1;
+  overflow: auto;
+  flex-wrap: wrap;
+  color: white;
+  align-items: center;
+  justify-content: center;
 }
