+++ src/components/burger-ingredients/burger-ingredients.module.css
@@ -1,3 +1,14 @@
 .section {
-  flex-basis: 50%;
+  display: flex;
+  flex: 1;
+  overflow: auto;
+  flex-wrap: wrap;
+  color: white;
+  align-items: center;
+  justify-content: center;
+}
+
+.container {
+  overflow: auto;
+  height: 100%;
 }
