+++ src/components/burger-ingredients/ingredients-type/ingredients-type.module.css
@@ -1,0 +1,4 @@
+.container {
+  display: grid;
+  grid-template-columns: 1fr 1fr;
+}
